requirements.txt
azure-core==1.20.1
azure-data-tables==12.1.0
requests==2.22.0
